daprApps_v1/video-sharing/video-scale/test/requirements.txt
dapr-ext-grpc-dev >= 1.5.0rc1.dev
dapr-dev >= 1.5.0rc1.dev
ffmpeg-python